# docs/specification/openapi.yaml
openapi: 3.1.0
info:
  title: TestModeller API
  version: 0.1.0
  description: >
    Contract between the Angular UI and the Rust backend. Changes require explicit approval (see AGENTS.md).
    Hierarchy: Project > Component > Feature > Model. A Feature has a scenario description and one or more Models
    (state machines). Test cases belong to a Feature and are assigned to Model states or transitions (steps).
servers:
  - url: /api/v1
...
paths:
...
  /ai/proposals:
    post:
      tags: [AI]
      summary: Request AI proposals. Runs asynchronously; poll the job.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ProposalRequest' } } } }
      responses:
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } }
        '503': { $ref: '#/components/responses/Problem' }
  /jobs/{jobId}:
    parameters: [ { name: jobId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } } ]
    get:
      tags: [AI]
      summary: Job status; result lists created proposal ids
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } }
  /projects/{projectId}/proposals:
    parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
    get:
      tags: [AI]
      summary: List proposals
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ProposalStatus' } }
        - { name: featureId, in: query, schema: { $ref: '#/components/schemas/Id' } }
        - { name: modelId, in: query, schema: { $ref: '#/components/schemas/Id' } }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ProposalPage' } } } }
  /proposals/{proposalId}:
    parameters: [ { $ref: '#/components/parameters/ProposalId' } ]
    get:
      tags: [AI]
      summary: Get proposal
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Proposal' } } } }
  /proposals/{proposalId}/accept:
    parameters: [ { $ref: '#/components/parameters/ProposalId' } ]
    post:
      tags: [AI]
      summary: Accept proposal, optionally with an edited payload. Creates the entity with origin=ai.
      requestBody:
        content: { application/json: { schema: { type: object, properties: { payload: { type: object } } } } }
      responses:
        '200': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/Proposal' } } } }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /proposals/{proposalId}/reject:
    parameters: [ { $ref: '#/components/parameters/ProposalId' } ]
    post:
      tags: [AI]
      summary: Reject proposal
      requestBody:
        content: { application/json: { schema: { type: object, properties: { reason: { type: string } } } } }
      responses:
        '200': { description: Rejected, content: { application/json: { schema: { $ref: '#/components/schemas/Proposal' } } } }
  /settings/ai:
    get:
      tags: [Settings]
      summary: Get AI provider settings (secrets are never returned)
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AiSettings' } } } }
    put:
      tags: [Settings]
      summary: Update AI provider settings
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AiSettingsInput' } } } }
      responses:
        '200': { description: Saved, content: { application/json: { schema: { $ref: '#/components/schemas/AiSettings' } } } }
components:
  parameters:
    ProjectId: { name: projectId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
...
    ProposalId: { name: proposalId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
...
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
    Cursor: { name: cursor, in: query, schema: { type: string } }
...
  responses:
    Problem:
      description: Error (RFC 7807)
      content:
        application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
  schemas:
    Id: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time }
...
    ProposalStatus: { type: string, enum: [pending, accepted, rejected] }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        errors:
          type: array
          items: { type: object, properties: { field: { type: string }, message: { type: string } } }
...
    ProposalKind: { type: string, enum: [model, states-and-transitions, test-cases, feature-description] }
    ProposalRequest:
      type: object
      required: [kind, featureId]
      properties:
        kind: { $ref: '#/components/schemas/ProposalKind' }
        featureId: { $ref: '#/components/schemas/Id' }
        modelId: { $ref: '#/components/schemas/Id' }
        prompt: { type: string, maxLength: 8000, description: Free-text instruction from the user }
        count: { type: integer, minimum: 1, maximum: 20 }
        previewContext: { type: boolean, default: false, description: Return the context that would be sent instead of calling the model }
    Proposal:
      type: object
      required: [id, projectId, kind, status, payload, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        projectId: { $ref: '#/components/schemas/Id' }
        featureId: { $ref: '#/components/schemas/Id' }
        modelId: { $ref: '#/components/schemas/Id' }
        kind: { $ref: '#/components/schemas/ProposalKind' }
        status: { $ref: '#/components/schemas/ProposalStatus' }
        payload: { type: object, description: Shape matches the target entity input (ModelInput, StateInput/TransitionInput lists, TestCaseInput list) }
        rationale: { type: string }
        source: { type: string, description: Model id of the LLM }
        resultingEntityId: { $ref: '#/components/schemas/Id' }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
    ProposalPage:
      type: object
      properties: { items: { type: array, items: { $ref: '#/components/schemas/Proposal' } }, nextCursor: { type: [string, 'null'] } }
    Job:
      type: object
      required: [id, status]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        status: { type: string, enum: [queued, running, succeeded, failed] }
        proposalIds: { type: array, items: { $ref: '#/components/schemas/Id' } }
        error: { $ref: '#/components/schemas/Problem' }
    AiSettings:
      type: object
      properties:
        provider: { type: string, enum: [openai-compatible, anthropic, local, none] }
        baseUrl: { type: string, format: uri }
        model: { type: string }
        secretConfigured: { type: boolean }
        maxTokensPerRequest: { type: integer }
    AiSettingsInput:
      allOf:
        - { $ref: '#/components/schemas/AiSettings' }
        - type: object
          properties: { apiKey: { type: string, writeOnly: true } }